# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/colors": {
            "get": {
                "description": "Get a list of Color.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Color"
                ],
                "summary": "Get all Color.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Color"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Color.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Color"
                ],
                "summary": "Create New Color.",
                "parameters": [
                    {
                        "description": "the body to create a new Color",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ColorInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Color"
                        }
                    }
                }
            }
        },
        "/colors/{id}": {
            "get": {
                "description": "Get an Color by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Color"
                ],
                "summary": "Get Color.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Color id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Color"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Color by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Color"
                ],
                "summary": "Delete one Color.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Color id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Color by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Color"
                ],
                "summary": "Update Color.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Color id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age rating category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ColorInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Color"
                        }
                    }
                }
            }
        },
        "/colors/{id}/phones": {
            "get": {
                "description": "Get all Phones by ColorId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Color"
                ],
                "summary": "Get Phones.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Color id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Phone"
                            }
                        }
                    }
                }
            }
        },
...
        "/models": {
            "get": {
                "description": "Get a list of Model.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "Get all Model.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Model"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Model.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "Create New Model.",
                "parameters": [
                    {
                        "description": "the body to create a new Model",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ModelInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Model"
                        }
                    }
                }
            }
        },
        "/models/{id}": {
            "get": {
                "description": "Get an Model by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "Get Model.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Model"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Model by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "Delete one Model.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Model by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "Update Model.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age rating category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ModelInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Model"
                        }
                    }
                }
            }
        },
        "/models/{id}/phones": {
            "get": {
                "description": "Get all Phones by ModelId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "Get Phones.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Phone"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.ColorInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "controller.ModelInput": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Color": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Model": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Phone": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "integer"
                },
                "color_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "model_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "string"
                },
                "storage": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}